Styling Added - See Notes
I know the header says net on the search for both sides.
I have an issue with the results and how they're displayed presently and it will take me a bit of time to work out the kinks. Stick with me, I will have it taken care of by the end of Wednesday night.